+++ uni_modules/uview-ui/components/u-navbar/u-navbar.vue
@@ -46,7 +46,8 @@
 					<text
 						class="u-line-1 u-navbar__content__title"
 						:style="{
-							width: $u.addUnit(titleWidth)
+							width: $u.addUnit(titleWidth),
+							...$u.addStyle(titleStyle)
 						}"
 					>{{ title }}</text>
 				</slot>
@@ -93,6 +94,7 @@
 	 * @property {String | Number}	leftIconSize		左侧返回图标的大小（默认 20px ）
 	 * @property {String | Number}	leftIconColor		左侧返回图标的颜色（默认 #303133 ）
 	 * @property {Boolean}	        autoBack			点击左侧区域(返回图标)，是否自动返回上一页（默认 false ）
+	 * @property {Object | String}	titleStyle			标题的样式，对象或字符串
 	 * @event {Function} leftClick		点击左侧区域
 	 * @event {Function} rightClick		点击右侧区域
 	 * @example <u-navbar title="剑未配妥，出门已是江湖" left-text="返回" right-text="帮助" @click-left="onClickBack" @click-right="onClickRight"></u-navbar>