+++ uni_modules/uview-ui/components/u-number-box/u-number-box.vue
@@ -134,8 +134,7 @@
 		computed: {
 			getCursorSpacing() {
 				// 判断传入的单位，如果为px单位，需要转成px
-				const number = parseInt(this.cursorSpacing)
-				return /rpx$/.test(String(this.cursorSpacing)) ? uni.upx2px(number) : number
+				return uni.$u.getPx(this.cursorSpacing)
 			},
 			// 按钮的样式
 			buttonStyle() {