+++ uni_modules/uview-ui/components/u-picker/u-picker.vue
@@ -18,6 +18,7 @@
 				class="u-picker__view"
 				:indicatorStyle="`height: ${$u.addUnit(itemHeight)}`"
 				:value="innerIndex"
+				:immediateChange="immediateChange"
 				:style="{
 					height: `${$u.addUnit(visibleItemCount * itemHeight)}`
 				}"
@@ -70,6 +71,7 @@
  * @property {String}			keyName				选项对象中，需要展示的属性键名（默认 'text' ）
  * @property {Boolean}			closeOnClickOverlay	是否允许点击遮罩关闭选择器（默认 false ）
  * @property {Array}			defaultIndex		各列的默认索引
+ * @property {Boolean}			immediateChange		是否在手指松开时立即触发change事件（默认 false ）
  * @event {Function} close		关闭选择器时触发
  * @event {Function} cancel		点击取消按钮触发
  * @event {Function} change		当选择值变化时触发