+++ uni_modules/uview-ui/components/u-popup/u-popup.vue
@@ -4,7 +4,7 @@
 			:show="show"
 			@click="overlayClick"
 			v-if="overlay"
-			:duration="duration"
+			:duration="overlayDuration"
 			:customStyle="overlayStyle"
 			:opacity="overlayOpacity"
 		></u-overlay>
@@ -12,7 +12,7 @@
 			:show="show"
 			:customStyle="transitionStyle"
 			:mode="position"
-			:duration="duration"
+			:duration="duration1"
 			@after-enter="afterEnter"
 			@click="clickHandler"
 		>
@@ -55,6 +55,7 @@
 	 * @property {Boolean}			overlay				是否显示遮罩 （默认 true ）
 	 * @property {String}			mode				弹出方向（默认 'bottom' ）
 	 * @property {String | Number}	duration			动画时长，单位ms （默认 300 ）
+	 * @property {String | Number}	overlayDuration			遮罩层动画时长，单位ms （默认 350 ）
 	 * @property {Boolean}			closeable			是否显示关闭图标（默认 false ）
 	 * @property {Object | String}	overlayStyle		自定义遮罩的样式
 	 * @property {String | Number}	overlayOpacity		遮罩透明度，0-1之间（默认 0.5）
@@ -74,7 +75,9 @@
 		name: 'u-popup',
 		mixins: [uni.$u.mpMixin, uni.$u.mixin, props],
 		data() {
-			return {}
+			return {
+				overlayDuration: this.duration + 50
+			}
 		},
 		watch: {
 			show(newValue, oldValue) {