+++ uni_modules/uview-ui/libs/config/config.js
@@ -1,5 +1,5 @@
-// 此版本发布于2022-01-14
-const version = '2.0.20'
+// 此版本发布于2022-01-20
+const version = '2.0.22'
 
 // 开发环境才提示，生产环境不会提示
 if (process.env.NODE_ENV === 'development') {