+++ uni_modules/uview-ui/changelog.md
@@ -1,3 +1,21 @@
+## 2.0.22（2022-01-19）
+# uView2.0重磅发布，利剑出鞘，一统江湖
+
+1. $u.page()方法优化，避免在特殊场景可能报错的问题
+2. picker组件添加immediateChange参数
+3. 新增$u.pages()方法
+## 2.0.21（2022-01-19）
+# uView2.0重磅发布，利剑出鞘，一统江湖
+
+1. 优化：form组件在用户设置rules的时候提示用户model必传
+2. 优化遗留的通过正则判断rpx单位的问题
+3. 修复微信小程序环境中tabbar组件开启safeAreaInsetBottom属性后，placeholder高度填充不正确
+4. 修复swiper在current指定非0时缩放有误
+5. 修复u-icon添加stop属性的时候报错
+6. 修复upload组件在accept=all的时候没有作用
+7. 修复在text组件mode为phone时call属性无效的问题
+8. 处理u-form clearValidate方法
+9. 其他修复
 ## 2.0.20（2022-01-14）
 # uView2.0重磅发布，利剑出鞘，一统江湖